edited the format of the files
