+++ config.json
@@ -1,11 +1,11 @@
 {
-  "bot_settings": {
-    "command_channels": [
-      1192895746672951340
-    ],
-    "ta_review_channel": [
-      1284224818698260490
-    ],
+  "quest": {
+    "state_path": "state/quest/"
+  },
+  "metrics": {
+    "metrics_path": "state/metrics/"
+  },
+  "rubber_duck": {
     "channels": [
       {
         "name": "duck-pond",
@@ -14,7 +14,7 @@
       {
         "name": "other-pond",
         "prompt_file": "prompts/standard-rubber-duck.txt",
-        "engine": "gpt-3.5-turbo-16k"
+        "engine": "gpt-4o-mini"
       },
       {
         "name": "gpt4",
@@ -30,19 +30,37 @@
         "timeout": 10
       }
     ],
-    "admin_ids": [
-      933123843038535741
-    ],
     "defaults": {
       "engine": "gpt-4o",
       "timeout": 600
     }
   },
-  "duck_settings": {
-    "retry_protocol": {
-      "max_retries": 2,
-      "delay": 2,
-      "backoff": 2
+  "feedback": {
+    "cs101": {
+      "ta_feedback_channel": 12871096109832741098,
+      "ta_role_id": 32742987340293740298347
+    },
+    "cs111": {
+      "ta_feedback-channel": 222234234234,
+      "ta_role_id": 4356356712345684
+    },
+    "cs260": {
+      "ta_feedback-channel": 2223423241234,
+      "ta_role_id": 14352136312356745684
+    },
+    "cs312": {
+      "ta_feedback-channel": 123422234234,
+      "reviewer_role_id": 56412356356745684
     }
+  },
+  "open_ai_retry_protocol": {
+    "max_retries": 2,
+    "delay": 2,
+    "backoff": 2
+  },
+  "admin_settings": {
+    "admin_channel_id": 9817278904209783,
+    "admin_role_id": 123892344230
   }
 }
+
