--- config_proposal.json
@@ -1,61 +1,0 @@
-{
-  "storage_settings": {
-    "state_path": "state_path"
-  },
-
-  "bot_settings": {
-    "admin": {
-      "command_channels": [
-        1192895746672951340
-      ],
-      "admin_ids": [
-        933123843038535741
-      ]
-    },
-    "duck_settings": {
-      "channels": [
-        {
-          "name": "duck-pond",
-          "prompt_file": "prompts/standard-rubber-duck.txt"
-        },
-        {
-          "name": "other-pond",
-          "prompt_file": "prompts/standard-rubber-duck.txt",
-          "engine": "gpt-4o-mini"
-        },
-        {
-          "name": "gpt4",
-          "engine": "gpt-4o"
-        },
-        {
-          "name": "gpt-3.5-turbo",
-          "engine": "gpt-3.5-turbo"
-        },
-        {
-          "name": "fast",
-          "engine": "gpt-3.5-turbo",
-          "timeout": 10
-        }
-      ],
-      "defaults": {
-        "engine": "gpt-4o",
-        "timeout": 600
-      },
-      "retry_protocol": {
-        "max_retries": 2,
-        "delay": 2,
-        "backoff": 2
-      },
-      "feedback": {
-        "cs1101232342345346265": {
-          "ta-feedback-channel": 12871096109832741098,
-          "ta-role-id": 32742987340293740298347
-        },
-        "cs2345u40528340958": {
-          "ta-feedback-channel": 22234234,
-          "ta-role-id": 4356356745684
-        }
-      }
-    }
-  }
-}
