+++ rubber_duck/rubber_duck.py
